Run the garbage collection synchronously
Running it asynchronously might cause issues, and finalizers are usually
quite fast.

One particular issue we've been seeing is with SQLite.  Because of the
behavior of the POSIX Advisory Locks, it can't free file descriptors if
there are other locks held on that inode by the same process.

If we don't allow the API user to control when the finalizers run, they
can run after the database has already been opened again, thus impeding
SQLite from reusing or freeing that file descriptor.